refactor(yank): decrease the timer for `highlight`
I got to know that short timer is very fast for the `TextYankPost`
event. It really feels fast. I also have noticed in my workflow that I
tend to wait for that highlight to go off. That's why decreasing the
timer from `500` in `yanky.nvim` and `400` in my `autcmd` to `40`.