refactor(plugins): move plugins to `after/plugin`
It seems that `tjdevries` is using `after/plugin` director for his almost all
the plugins. So moving all my plugins which are not creating any problem
to `after/plugin` directory. That seems reasonable.

Removing `auto-session` & `session-lens` because it was messing up the
startuptime. I have `harpoon` and `telescope` to go to any file in the
project very fast.