refactor(plugins): move plugins to `after/plugin`
moving plugins to `after/plugin` directory becuase `ThePrimeagen` and
`tjdevries` are doing the same thing and it also seems reasonable.

remove `auto-session` and `session-lens` because it was increasing my
`startuptime`. Anyways I don't need these because I have
`telescope.nvim` and `harpoon` to go to any file in the project asap.